move file from model to models folder
