+++ Realta.Contract/Models/StockDetailDto.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Realta.Contract
+namespace Realta.Contract.Models
 {
     public class StockDetailDto
     {
