+++ Realta.Contract/Models/StockPhotoGroupDto.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Realta.Contract.Model
+namespace Realta.Contract.Models
 {
     public class StockPhotoGroupDto
     {
